Renamed to lessontemplate
