+++ src/main/resources/lessons/lessontemplate/db/migration/V2019_11_10_1__introduction.sql
