+++ src/main/resources/lessons/lessontemplate/documentation/lesson-template-attack.adoc
