+++ src/main/resources/lessons/lessontemplate/documentation/lesson-template-content.adoc
