+++ src/main/resources/lessons/lessontemplate/documentation/lesson-template-database.adoc
