+++ src/main/resources/lessons/lessontemplate/documentation/lesson-template-intro.adoc
