+++ src/main/resources/lessons/lessontemplate/documentation/lesson-template-lesson-class.adoc
