+++ src/main/resources/lessons/lessontemplate/documentation/lesson-template-video.adoc
