+++ src/main/resources/lessons/lessontemplate/i18n/WebGoatLabels.properties
