+++ src/main/resources/lessons/lessontemplate/js/idor.js
